write jpnb and py
